Fixing Mozilla Science Lab logo in "About Us"
The svg logo of Mozilla Science Lab use some special font that can
be unavailable in some machines and for that reason we should use
a png version of it.

Similar problem had been report in #250 and fix in #254 for home
page and now we fix it to "About Us".