_includes/projects/ipfs.html: Add a project blurb for IPFS
This is a fun side project I've been working on. You can use it as the
base layer for lots of different tools, but the most relevant to the
SWC community are likely:

* Git-style version control with peer-to-peer content distribution.
* Archive, version, and share your huge, possibly binary, datasets
  easily and efficiently.

The basic architecture will be is familiar to Git users
(content-addressable Merkle graphs!) and BitTorrent users
(peer-to-peer data distribution for fast, robust sharing!).